+++ anagix_init/metadata.rb
@@ -4,6 +4,6 @@ maintainer_email "seijiro.moriyama@anagix.com"
 license          "All rights reserved"
 description      "Installs/Configures anagix_init"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.2.0"
-depends 'apt'
-depends 'build-essential'
+version          "0.3.0"
+#depends 'apt'
+#depends 'build-essential'